% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infographics.R
\name{fct_calc_indic_evol}
\alias{fct_calc_indic_evol}
\title{Calcul des indicateurs traceur d'évolution}
\usage{
fct_calc_indic_evol(RPU, group_by = "etab", excl_orient_non_pec = TRUE)
}
\arguments{
\item{RPU}{base RPU avec variables intermédiaires déjà calculées}

\item{group_by}{Façon dont sont regroupées les résultats : "none" pour population globale, "etab" pour groupe par étab et "group" pour grouper par "group" de taille d'étab}

\item{excl_orient_non_pec}{Booléen : Les fugues, réorientations, partis sans attendre et sorties contre avis doivent-il être exclus des délais et diag ?}
}
\value{
Un tibble
}
\description{
Calcul des indicateurs traceur d'évolution
}
