% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconQual.R
\name{fct_create_icon_qual_folder}
\alias{fct_create_icon_qual_folder}
\title{Créer dossier icon qualité}
\usage{
fct_create_icon_qual_folder(path)
}
\arguments{
\item{path}{Chemin vers le dossier ou créer le dossier "iconQualiteDonnee et les 100 icons}
}
\description{
Créer dossier icon qualité
}
