% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagCharge.R
\name{fct_diagramme_charge}
\alias{fct_diagramme_charge}
\title{Figure et table diag de charge}
\usage{
fct_diagramme_charge(
  base_etab,
  base_groupe,
  borne_min,
  borne_max,
  duree_max,
  pedia
)
}
\arguments{
\item{base_etab}{RPU de l'établissement}

\item{base_groupe}{RPU du groupe}

\item{borne_min}{Début de la période}

\item{borne_max}{Fin de la période}

\item{duree_max}{Durée maximale autorisée}

\item{pedia}{booléen : rapport pédia ou non}
}
\value{
un plot et 2 table et 2 infos pour l'infographie
}
\description{
Figure et table diag de charge
}
