% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delais.R
\name{fct_duree_passage}
\alias{fct_duree_passage}
\title{Lollipop durées de passages}
\usage{
fct_duree_passage(base_groupe, titre_tab, excl_orient_non_pec)
}
\arguments{
\item{base_groupe}{RPU du groupe}

\item{titre_tab}{Titre du tableau}

\item{excl_orient_non_pec}{Booléen : Les patients réorientés/fugues/sortie sans ou contre avis médical doivent-ils être exclus ?}
}
\value{
une liste
}
\description{
Lollipop durées de passages
}
