% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fct_format_control}
\alias{fct_format_control}
\title{Controle les modalités d'une variable}
\usage{
fct_format_control(var, val_autor, lab)
}
\arguments{
\item{var}{Vecteur contenant la variable à controler}

\item{val_autor}{Vecteur des valeurs autorisées}

\item{lab}{Nom de la variable à imprimer dans le retour d'info}
}
\value{
vecteur contenant var corrigé en "NC" pour les données Non Conformes
}
\description{
Controle les modalités d'une variable
}
