% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infographics.R
\name{fct_infographic_descript_PA}
\alias{fct_infographic_descript_PA}
\title{Inforgraphie description des personnes âgées}
\usage{
fct_infographic_descript_PA(RPU, nb_PA_region, lab_annee = NULL)
}
\arguments{
\item{RPU}{Une table au format RPU utilisé dans les rapports (non filtré sur les PA)}

\item{nb_PA_region}{Nombre de personnes âgées en ARA à afficher dans le graphique}

\item{lab_annee}{label des années à afficher. Calculé automatiquement sur les données si "NULL"}
}
\value{
Un ggplot
}
\description{
Inforgraphie description des personnes âgées
}
