% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infographics.R
\name{fct_infographic_evol}
\alias{fct_infographic_evol}
\title{Inforgraphie évolution}
\usage{
fct_infographic_evol(RPU_evol)
}
\arguments{
\item{RPU_evol}{Table d'évolition entre l'année n-1 et n. Créé par la fonction fct_make_tab_evol_RPU}
}
\value{
Un ggplot
}
\description{
Inforgraphie évolution
}
