% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infographics.R
\name{fct_infographic_resume}
\alias{fct_infographic_resume}
\title{Infographic nid d'abeille}
\usage{
fct_infographic_resume(base, pediatrique = FALSE, tab_activite, type)
}
\arguments{
\item{base}{RPU (établissement ou groupe)}

\item{pediatrique}{Booléen : Rapport pédiatrique T/F}

\item{tab_activite}{Valeur retournée par la fonction}

\item{type}{character : Etab ou Groupe selon si la base fournie est celle du groupe ou de l'étab}
}
\value{
Un graphique
}
\description{
Infographic nid d'abeille
}
