% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPlot.R
\name{fct_make_barfill_par_etab}
\alias{fct_make_barfill_par_etab}
\title{barplot position fill par etablissement}
\usage{
fct_make_barfill_par_etab(
  base_groupe,
  var_interet,
  label_var_interet,
  grouping_var,
  label_grouping,
  titre = "",
  make_table = FALSE,
  na.rm = FALSE,
  pedia,
  excl_orient_non_pec = F,
  palette_color_manual = NULL,
  anonyme = TRUE
)
}
\arguments{
\item{base_groupe}{RPU du groupe de l'établissement}

\item{var_interet}{Character : Variable a représenter}

\item{label_var_interet}{Nom de la variable à représenter}

\item{grouping_var}{Nom de la variable de groupement (en axe y)}

\item{label_grouping}{Label a afficher pour l'axe Y sur le graphique}

\item{titre}{Titre du graphique}

\item{make_table}{Booléen : Le tableau de données doit-il être généré}

\item{na.rm}{Booléen : Retirer les données manquantes}

\item{pedia}{Booléen : le rapport est-il pédiatrique}

\item{excl_orient_non_pec}{Booléen : Les patients réorientés/fugues/sortie sans ou contre avis médical doivent-ils être exclus ?}

\item{palette_color_manual}{palette de couleur donnée à "value" dans "scale_fill_manual"}

\item{anonyme}{Boolèen : Les valeurs de la variable de groupement doivent-elles être cachées sur le graphique ?}
}
\value{
un plot et deux tables
}
\description{
barplot position fill par etablissement
}
