% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPlot.R
\name{fct_make_barplot}
\alias{fct_make_barplot}
\title{barplot simple}
\usage{
fct_make_barplot(
  base_groupe,
  var_interet,
  ylab = "",
  titre_graph = "",
  titre_tab = "",
  lab_var_tab = ""
)
}
\arguments{
\item{base_groupe}{RPU du groupe dont l'étab actif (var etab_actif)}

\item{var_interet}{Character : Variable a représenter}

\item{ylab}{Nom de l'axe y du graphique}

\item{titre_graph}{Titre du graphique}

\item{titre_tab}{Titre de la table}

\item{lab_var_tab}{Nom de la variable dans la table}
}
\value{
Un plot et une table
}
\description{
barplot simple
}
