% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPlot.R
\name{fct_make_barplot_age}
\alias{fct_make_barplot_age}
\title{barplot simple sur l'âge}
\usage{
fct_make_barplot_age(base_groupe, titre_tab = "")
}
\arguments{
\item{base_groupe}{RPU du groupe dont l'étab actif (var etab_actif)}

\item{titre_tab}{titre de la table}
}
\value{
Un plot et une table
}
\description{
barplot simple sur l'âge
}
