% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infographics.R
\name{fct_make_tab_evol_RPU}
\alias{fct_make_tab_evol_RPU}
\title{table evolution infographie}
\usage{
fct_make_tab_evol_RPU(
  RPU_n1,
  RPU_n2,
  group_by = "etab",
  excl_orient_non_pec = TRUE
)
}
\arguments{
\item{RPU_n1}{RPU de l'année de référence}

\item{RPU_n2}{RPU de l'année "active" du rapport}

\item{group_by}{Façon dont sont regroupées les résultats : "none" pour population globale, "etab" pour groupe par étab et "group" pour grouper par "group" de taille d'étab}

\item{excl_orient_non_pec}{Booléen : Les fugues, réorientations, partis sans attendre et sorties contre avis doivent-il être exclus des délais et diag ?}
}
\value{
un tibble
}
\description{
table evolution infographie
}
