% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fct_min_as_hour}
\alias{fct_min_as_hour}
\title{Transformation minutes en heures}
\usage{
fct_min_as_hour(d_min)
}
\arguments{
\item{d_min}{numeric}
}
\value{
Retourne une durée en minute au format charactère HH:MM
}
\description{
Transformation minutes en heures
}
