% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fct_nb_presents}
\alias{fct_nb_presents}
\title{Nombre patients presents une heure données}
\usage{
fct_nb_presents(d_entree, d_sortie, heure, avg_etab = FALSE, id_etab = NULL)
}
\arguments{
\item{d_entree}{POSIXct date/heure d'entrée}

\item{d_sortie}{POSIXct date/heure d'entrée}

\item{heure}{Heure à tester (Entier entre 0 à 23)}

\item{avg_etab}{Booléen : Le nombre de présent doit-il être au total ou en moyenne par établissement}

\item{id_etab}{character : Vecteur contenant un identifiant unique des établissements présents}
}
\value{
numeric
}
\description{
Nombre patients presents une heure données
}
