% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramideAge.R
\name{fct_pyramide_age}
\alias{fct_pyramide_age}
\title{Pyramide des ages}
\usage{
fct_pyramide_age(base, titre_tab, bassin_calc)
}
\arguments{
\item{base}{base contenant les variables suivantes (SEXE, age_class, base, n). Age_classe contient les valeurs suivantes c("00_02", "03_10", "11_17", "18_24", "25_39", "40_54", "55_64", "65_79", "80_P")}

\item{titre_tab}{Titre du tableau à retourner}

\item{bassin_calc}{Booléen indiquant si le bassin sanitaire était calculable}
}
\value{
un plot et une table
}
\description{
Pyramide des ages
}
