% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fct_tab_diag}
\alias{fct_tab_diag}
\title{tableau des diagnostics principaux}
\usage{
fct_tab_diag(base_etab, base_groupe, excl_orient_non_pec = FALSE)
}
\arguments{
\item{base_etab}{RPU de l'établissement}

\item{base_groupe}{RPU du groupe de l'établissement}

\item{excl_orient_non_pec}{Booléen : Les patients réorientés/fugues/sortie sans ou contre avis médical doivent-ils être exclus ?}
}
\value{
Une liste
}
\description{
tableau des diagnostics principaux
}
