% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tab_CIM10_eCerveau}
\alias{tab_CIM10_eCerveau}
\title{Thésaurus des diagnostics SARA}
\format{
\subsection{\code{tab_CIM10_eCerveau}}{

Un tibble de 42 883 lignes et 11 colonnes
\describe{
\item{code_diag}{Code Cim10}
}
}
}
\usage{
tab_CIM10_eCerveau
}
\description{
Thésaurus utilisé par SARA dans eCerveau
}
\keyword{datasets}
