% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Thesaurus_Fedoru}
\alias{Thesaurus_Fedoru}
\title{Thésaurus des diagnostics de la FEDORU}
\format{
\subsection{\code{Thesaurus_Fedoru}}{

Un tibble de 22 567 lignes et 5 colonnes
\describe{
\item{diagnostic_code}{Code Cim10}
\item{diagnostic_libelle}{Diagnostic associé au code cim10}
\item{type_urgence_libelle}{Classification FEDORU du diagnostic}
\item{discipline_libelle}{Discipline associée au diagnostic}
\item{pathologie_libelle}{Pathologie associée au diagnostic}
}
}
}
\usage{
Thesaurus_Fedoru
}
\description{
Thésaurus utilisé dans le kit panorama 2024
}
\keyword{datasets}
