% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fct_jpeg_ggplot}
\alias{fct_jpeg_ggplot}
\title{Convertir jpeg en ggplot}
\usage{
fct_jpeg_ggplot(path, compression = "50\%")
}
\arguments{
\item{path}{Chemin vers le fichier jpeg}

\item{compression}{char : Argument geometry utilisé dans magick::image_resize. geometry = "50\%" par défaut}
}
\value{
ggplot
}
\description{
Utilisée pour convertir une image en ggplot afin de pouvoir ajouter le logo + macaron qualité de la données
}
