% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageGarde.R
\name{fct_pageGarde}
\alias{fct_pageGarde}
\title{Page de garde}
\usage{
fct_pageGarde(nom_etab, date_debut, date_fin)
}
\arguments{
\item{nom_etab}{character : Nom de l'établissement}

\item{date_debut}{character DMY : date de début du rapport}

\item{date_fin}{character DMY : date de fin du rapport}
}
\value{
un plot
}
\description{
Génère la page de garde des rapports RPU. ATTENTION : La police utilisée ne fonctionne que dans un environnement utilisant showtext::showtext_auto
}
